use dedicated aral for replication sender

+++ src/streaming/stream-replication-sender.h
@@ -31,7 +31,7 @@ void replication_sender_delete_pending_requests(struct sender_state *sender);
 void replication_sender_request_add(struct sender_state *sender, const char *chart_id, time_t after, time_t before, bool start_streaming);
 void replication_sender_recalculate_buffer_used_ratio_unsafe(struct sender_state *s);
 
-size_t replication_sender_allocated_memory(void);
+int64_t replication_sender_allocated_memory(void);
 size_t replication_sender_allocated_buffers(void);
 
 int replication_prefetch_default(void);
